renamed all components to proper naming schemes -- avoiding index.tsx
