+++ client/src/components/PairedRoom/PairedRoom.tsx
@@ -6,10 +6,10 @@ import useAgoraClient from '../../hooks/Agora/useAgoraClient';
 import useSocketListeners from '../../hooks/Sockets/useSocketListeners';
 import useFetchToken from '../../hooks/Agora/useFetchToken';
 import StartCall from './StartCall';
-import PairedChat from './PairedChat/Chat';
-import PairedVideos from './VideoCams/VideoCams';
-import Footer from '../Footer';
-import PairedContent from './PairedContent/Content';
+import Chat from './Chat/Chat';
+import VideoCams from './VideoCams/VideoCams';
+import Footer from '../Footer/Footer';
+import Content from './Content/Content';
 import './PairedRoom.css';
 
 const CodeCollab: React.FC = () => {
@@ -40,9 +40,9 @@ const CodeCollab: React.FC = () => {
                 <>
                     <main id="video-main-wrapper">
                         <AgoraRTCProvider client={agoraEngine}>
-                            <PairedVideos channelName={channelName} />
-                            <PairedContent agoraEngine={agoraEngine} leaveRoomHandler={leaveRoomHandler} />
-                            <PairedChat channelName={channelName} agoraEngine={agoraEngine} />
+                            <VideoCams channelName={channelName} />
+                            <Content agoraEngine={agoraEngine} leaveRoomHandler={leaveRoomHandler} />
+                            <Chat channelName={channelName} agoraEngine={agoraEngine} />
                         </AgoraRTCProvider>
                     </main>
                     <Footer />
